riscv: Use optimized memcpy,memcmp,memove,memset,strcmp,strcpy,strlen with -Os
In upstream code, building libc with -Os will disable the
optimized versions of these (falling back to byte-oriented versions).

Removing -Os and setting -O2 increases the total libc size by a large
amount (+22KB for a ROM build), but just enabling these is only +1KB or so.

This also brings memcpy() behaviour in sync with our Xtensa newlib builds, which
use the optimized versions of these functions even with -Os